+++ crates/js-sys/src/lib.rs
@@ -4812,6 +4812,13 @@ pub mod WebAssembly {
         #[wasm_bindgen(constructor, js_namespace = WebAssembly, catch)]
         pub fn new(table_descriptor: &Object) -> Result<Table, JsValue>;
 
+        /// The `WebAssembly.Table()` constructor creates a new `Table` object
+        /// of the given size and element type.
+        ///
+        /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WebAssembly/Table)
+        #[wasm_bindgen(constructor, js_namespace = WebAssembly, catch)]
+        pub fn new_with_value(table_descriptor: &Object, value: JsValue) -> Result<Table, JsValue>;
+
         /// The length prototype property of the `WebAssembly.Table` object
         /// returns the length of the table, i.e. the number of elements in the
         /// table.
@@ -4834,6 +4841,18 @@ pub mod WebAssembly {
         /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WebAssembly/Table/grow)
         #[wasm_bindgen(method, catch, js_namespace = WebAssembly)]
         pub fn grow(this: &Table, additional_capacity: u32) -> Result<u32, JsValue>;
+
+        /// The `grow()` prototype method of the `WebAssembly.Table` object
+        /// increases the size of the `Table` instance by a specified number of
+        /// elements.
+        ///
+        /// [MDN documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/WebAssembly/Table/grow)
+        #[wasm_bindgen(method, catch, js_namespace = WebAssembly, js_name = grow)]
+        pub fn grow_with_value(
+            this: &Table,
+            additional_capacity: u32,
+            value: JsValue,
+        ) -> Result<u32, JsValue>;
 
         /// The `set()` prototype method of the `WebAssembly.Table` object mutates a
         /// reference stored at a given index to a different value.